+++ Client/Game/Enviroment/BiomeManagement/Biomes/Empty.cs
@@ -20,7 +20,7 @@ namespace VoxelValley.Client.Game.Enviroment.BiomeManagement.Biomes
             return VoxelManager.AirVoxel;
         }
 
-        internal override void GetFinishers(int worldX, int worldZ, ushort chunkX, ushort chunkZ, ushort height, ref ushort[,,] voxels)
+        internal override void GetFinishers(int worldX, int worldZ, ushort chunkX, ushort chunkZ, int height, ref ushort[,,] voxels)
         {
 
         }
